Limit mit cit spec to applicable cases
